requirements.txt
# AI Study Assistant - Dependencies

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0  # For SemanticChunker

# BM25 retrieval
rank-bm25>=0.2.2

# LLM providers (install as needed)
langchain-groq>=0.2.0
langchain-mistralai>=0.2.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# Embeddings
sentence-transformers>=3.0.0

# Vector database
chromadb>=0.5.0

# Document processing
pypdf>=4.0.0
unstructured>=0.15.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Streamlit frontend
streamlit>=1.32.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
